Split scaladoc completion related logic into another file.
Since Completions.scala has already really big,
try to split scaladoc related logic into another file.
